requirements.txt
# Core dependencies
pandas>=2.0.0
chardet>=5.0.0
fastmcp>=1.0.0
psutil>=5.9.0
openpyxl>=3.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Note: Chart.js is a frontend dependency and should be included
# via CDN or npm in your frontend setup